requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
channels==4.1.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==4.2.16
django-jazzmin==3.0.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
h11==0.14.0
idna==3.10
multidict==6.1.0
phonenumbers==8.13.47
pillow==11.0.0
propcache==0.2.0
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
twilio==9.3.4
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
yarl==1.15.4
